Moved logger/logger.py to lib/.
