+++ lib/__init__.py
@@ -5,11 +5,13 @@ from .options import check_train_options, check_test_options
 from .env import make_split_provider
 from .dataloader import create_dataloader
 from .framework import create_model
+from .logger import get_logger
 
 __all__ = [
             'check_train_options',
             'check_test_options',
             'make_split_provider',
             'create_dataloader',
-            'create_model'
+            'create_model',
+            'get_logger'
             ]
