+++ lib/optimizer.py
@@ -2,7 +2,6 @@
 # -*- coding: utf-8 -*-
 
 import torch.optim as optim
-from typing import List
 
 
 class Optimizer:
