--- logger/__init__.py
@@ -1,6 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-from .logger import get_logger
-
-__all__ = ['get_logger']
